Add documentation regarding credentials. Also moved the SigV4RequestSigner out of the internal folder.
