+++ src/SigV4RequestSigner.ts
@@ -1,5 +1,5 @@
-import { Credentials, QueryParams, RequestSigner } from '../SignalingClient';
-import { validateValueNonNil } from './utils';
+import { Credentials, QueryParams, RequestSigner } from './SignalingClient';
+import { validateValueNonNil } from './internal/utils';
 type Headers = { [header: string]: string };
 
 /**
